refactor(voice): extract ComposeBar dictation into a runes controller
The ~250 inlined lines (engine pick, Web Speech session, MediaRecorder
recording, Web-Audio interim preview, teardown) move verbatim into
createDictation() in src/lib/dictation.svelte.ts. The controller is the
single owner of the getVoiceStatus probe with an explicit ready promise
and a guarded autoStart(), so the iOS user-activation-window auto-start
no longer depends on .then registration order. Interim ticks now carry
mode=partial. Behaviour-identical otherwise; new browser test covers the
web-speech path incl. auto-start and the no-upload-on-submit/cancel
guarantee.
